Refactor SignalR sync libraries.
