--- Client.Sync.SignalR/Client.Sync.SignalR.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>$(DefaultTargetFramework)</TargetFramework>
-    <AssemblyName>Quibble.Client.Sync.SignalR</AssemblyName>
-    <RootNamespace>Quibble.Client.Sync.SignalR</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Components" Version="6.0.0-rc.2.21480.10" />
-    <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="6.0.0-rc.2.21480.10" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0-rc.2.21480.5" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Client.Sync\Client.Sync.csproj" />
-    <ProjectReference Include="..\Shared.Sync.SignalR\Shared.Sync.SignalR.csproj" />
-    <ProjectReference Include="..\Shared\Shared.csproj" />
-  </ItemGroup>
-
-</Project>
