--- Shared.Sync.SignalR/Shared.Sync.SignalR.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>$(DefaultTargetFramework)</TargetFramework>
-    <AssemblyName>Quibble.Shared.Sync.SignalR</AssemblyName>
-    <RootNamespace>Quibble.Shared.Sync.SignalR</RootNamespace>
-    <NeutralLanguage>en-GB</NeutralLanguage>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Shared\Shared.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <SupportedPlatform Include="browser" />
-  </ItemGroup>
-
-</Project>
